// src/app/interfaces/forms/bio-jardineria.ts
export interface BioJardineria {
  formId: number;
  typeNatureId: number;
  identification: number;
  name1: string;
  name2: string;
  lastName1: string;
  lastName2: string;
  email: string;
  phoneNumber: string;
  mobile: number;
  sendEmailNotification: false;
  aditionalInfo: string;
  politicId: number;
  recaptchaToken: string;
}